More Query Log spam deletion options
In addition to allowing the deletion of spammy IPs from the query log, also add the ability to delete spammy search queries from the log as well.

As well, add a scorched earth "all associated" option which takes every IP that searched for a specific query and purges all of those IPs, along with any *other* queries they've made, from the Query Log.