Improve latin accent matching, and other things...
MySQL and MariaDB do a good job of automating matching of single characters to their accented and non-accented equivalents. However, some multi-letter transliterations are missed, such as 'aa' for a-ring. Use a subset of the latin accent matching array to ensure these transliterations are searched for in the database.

The highlighting script previously did not take into account that the text passed to it had been run through htmlspecialchars(). Adjust the regular expression to account for this. This enables highlighting of terms in search results that contain (for example) apostrophes like "don't".

A few minor changes to crawler.php in prep for a larger commit regarding the avoidance of race conditions. Dual crawls are still being triggered in some cases, so I'm working on that.