Provide raw url encoded request
Provide the raw url encoded version of the request for use in link URLs in the template. This prevents, for example, plus signs (+) being interpreted as spaces in pagination links.

This changes the default search template! You will need to update your template, or revert it and edit it again if you've changed yours.