Show single term 'phrases' as +important
+Important search terms were lumped in with phrase matches since they counted for the same type of bonus. This meant that literal '+important' tokens were showing up in the search Query Log as '"important"'. Fix this, so they show up with the + symbol instead.

Note that the `strpos()` call here doesn't check for a zero (0) return value as opposed to `false` since it's impossible for a term to have a space as the first character due to `trim()`. If there's a space in the term/phrase, it will always be at position 1 or later.