Race condition updates in crawler.php
The idea for these updates is to make very sure a crawl is not currently running (or not currently allowed to run based on the timeout value) then set the `sp_crawling` value to zero (0).

At the vert last step before committing to start the crawl, re-check the `sp_crawling` value. If it's not still zero, then another crawl has started before this one, so we should exit the process and let the other process continue. If it's still zero, then this process is the first in line to crawl; the *very next step* is to set the `sp_crawling` value to the random number ID rolled by this process.

There are probably still ways to refine this, especially the checking to see if no crawler is currently running.